+++ cli/src/api/mod.rs
@@ -10,7 +10,9 @@ use phylum_types::types::group::{
 use phylum_types::types::job::{
     AllJobsStatusResponse, SubmitPackageRequest, SubmitPackageResponse,
 };
-use phylum_types::types::package::{PackageDescriptor, PackageSpecifier, PackageSubmitResponse};
+use phylum_types::types::package::{
+    PackageDescriptor, PackageDescriptorAndLockfile, PackageSpecifier, PackageSubmitResponse,
+};
 use phylum_types::types::project::{
     CreateProjectRequest, CreateProjectResponse, ProjectSummaryResponse,
 };
@@ -277,7 +279,7 @@ impl PhylumApi {
     /// Submit a new request to the system
     pub async fn submit_request(
         &self,
-        package_list: &[PackageDescriptor],
+        package_list: &[PackageDescriptorAndLockfile],
         project: ProjectId,
         label: Option<String>,
         group_name: Option<String>,
@@ -500,11 +502,17 @@ mod tests {
 
         let client = build_phylum_api(&mock_server).await?;
 
-        let pkg = PackageDescriptor {
+        let package_descriptor = PackageDescriptor {
             name: "react".to_string(),
             version: "16.13.1".to_string(),
             package_type: PackageType::Npm,
         };
+
+        let pkg = PackageDescriptorAndLockfile {
+            package_descriptor,
+            lockfile: Some("package-lock.json".to_owned()),
+        };
+
         let project_id = ProjectId::new_v4();
         let label = Some("mylabel".to_string());
         client.submit_request(&[pkg], project_id, label, None).await?;
@@ -530,11 +538,17 @@ mod tests {
 
         let client = build_phylum_api(&mock_server).await?;
 
-        let pkg = PackageDescriptor {
+        let package_descriptor = PackageDescriptor {
             name: "react".to_string(),
             version: "16.13.1".to_string(),
             package_type: PackageType::Npm,
         };
+
+        let pkg = PackageDescriptorAndLockfile {
+            package_descriptor,
+            lockfile: Some("package-lock.json".to_owned()),
+        };
+
         let project_id = ProjectId::new_v4();
         let label = Some("mylabel".to_string());
         client.submit_request(&[pkg], project_id, label, None).await?;