+++ cli/src/commands/parse.rs
@@ -1,11 +1,12 @@
 //! `phylum parse` command for lockfile parsing
 
 use std::path::{Path, PathBuf};
+use std::vec::IntoIter;
 use std::{fs, io};
 
 use anyhow::{anyhow, Context, Result};
 use phylum_lockfile::{LockfileFormat, Package, PackageVersion, Parse, ThirdPartyVersion};
-use phylum_types::types::package::PackageDescriptor;
+use phylum_types::types::package::{PackageDescriptor, PackageDescriptorAndLockfile};
 use walkdir::WalkDir;
 
 use crate::commands::{CommandResult, ExitCode};
@@ -15,6 +16,31 @@ pub struct ParsedLockfile {
     pub path: PathBuf,
     pub format: LockfileFormat,
     pub packages: Vec<PackageDescriptor>,
+}
+
+pub struct ParsedLockfileIterator {
+    path: PathBuf,
+    packages: IntoIter<PackageDescriptor>,
+}
+
+impl Iterator for ParsedLockfileIterator {
+    type Item = PackageDescriptorAndLockfile;
+
+    fn next(&mut self) -> Option<Self::Item> {
+        self.packages.next().map(|package_descriptor| PackageDescriptorAndLockfile {
+            package_descriptor,
+            lockfile: Some(self.path.to_string_lossy().into_owned()),
+        })
+    }
+}
+
+impl IntoIterator for ParsedLockfile {
+    type IntoIter = ParsedLockfileIterator;
+    type Item = PackageDescriptorAndLockfile;
+
+    fn into_iter(self) -> Self::IntoIter {
+        ParsedLockfileIterator { path: self.path, packages: self.packages.into_iter() }
+    }
 }
 
 pub fn lockfile_types(add_auto: bool) -> Vec<&'static str> {
@@ -31,10 +57,11 @@ pub fn lockfile_types(add_auto: bool) -> Vec<&'static str> {
 pub fn handle_parse(matches: &clap::ArgMatches) -> CommandResult {
     let lockfiles = config::lockfiles(matches, phylum_project::get_current_project().as_ref())?;
 
-    let mut pkgs: Vec<PackageDescriptor> = Vec::new();
+    let mut pkgs: Vec<PackageDescriptorAndLockfile> = Vec::new();
 
     for lockfile in lockfiles {
-        pkgs.extend(parse_lockfile(lockfile.path, Some(&lockfile.lockfile_type))?.packages);
+        let parsed_lockfile = parse_lockfile(lockfile.path, Some(&lockfile.lockfile_type))?;
+        pkgs.extend(parsed_lockfile.into_iter());
     }
 
     serde_json::to_writer_pretty(&mut io::stdout(), &pkgs)?;