+++ extensions/phylum.ts
@@ -5,6 +5,7 @@ export type Package = {
   name: string;
   version: string;
   type: string;
+  lockfile?: string;
 };
 
 export type Lockfile = {
@@ -110,7 +111,7 @@ export class PhylumApi {
    *
    * ```
    * [
-   *   { name: "accepts", version: "1.3.8", type: "npm" },
+   *   { name: "accepts", version: "1.3.8", type: "npm", lockfile: "/path/to/lockfile" },
    *   { name: "ms", version: "2.0.0", type: "npm" },
    *   { name: "negotiator", version: "0.6.3", type: "npm" },
    *   { name: "ms", version: "2.1.3", type: "npm" }
@@ -118,6 +119,7 @@ export class PhylumApi {
    * ```
    *
    * Accepted package types are "npm", "pypi", "maven", "rubygems", "nuget", "cargo", and "golang"
+   * The `lockfile` is an optional string used to represent the path to the lockfile.
    *
    * @param packages - List of packages to analyze
    * @param project - Project name. If undefined, the `.phylum_project` file will be used