+++ CHANGELOG.md
@@ -8,6 +8,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+- Support for NuGet's `packages.lock.json` lockfiles
+
 ### Changed
 - Switch documentation hosting from Readme.com to Docusaurus