+++ CHANGELOG.md
@@ -8,6 +8,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Changed
+- Include lockfile paths when analyzing projects
+
 ## [5.5.0] - 2023-07-18
 
 ### Added